+++ src/remix/remix-wizard.ts
@@ -4,11 +4,9 @@ import chalk from 'chalk';
 
 import {
   addSentryCliRc,
-  askForProjectSelection,
-  askForSelfHosted,
-  askForWizardLogin,
   confirmContinueEvenThoughNoGitRepo,
   ensurePackageIsInstalled,
+  getOrAskForProjectData,
   getPackageDotJson,
   installPackage,
   isUsingTypeScript,
@@ -54,15 +52,10 @@ async function runRemixWizardWithTelemetry(
   // We expect `@remix-run/dev` to be installed for every Remix project
   await ensurePackageIsInstalled(packageJson, '@remix-run/dev', 'Remix');
 
-  const { url: sentryUrl } = await askForSelfHosted(options.url);
-
-  const { projects, apiKeys } = await askForWizardLogin({
-    promoCode: options.promoCode,
-    url: sentryUrl,
-    platform: 'javascript-remix',
-  });
-
-  const selectedProject = await askForProjectSelection(projects);
+  const { selectedProject, authToken } = await getOrAskForProjectData(
+    options,
+    'javascript-remix',
+  );
 
   await traceStep('Install Sentry SDK', () =>
     installPackage({
@@ -77,7 +70,7 @@ async function runRemixWizardWithTelemetry(
   const isV2 = isRemixV2(remixConfig, packageJson);
 
   await addSentryCliRc(
-    apiKeys.token,
+    authToken,
     selectedProject.organization.slug,
     selectedProject.name,
   );